# services/api-gateway/internal/docs/openapi/openapi-organization.yaml
openapi: 3.0.3
info:
  title: CivicOS Organization Service
  version: 1.0.0
  description: |
    Organizations (registry + membership), announcements, projects, issue
    assignments ("receive reports"), and progress updates ("respond publicly",
    "update progress").

    All routes are proxied by the API gateway at `/api/v1/*`. The service
    itself mounts them under `/v1/*`.

    ## Authorization model

    Roles allowed to **create** a top-level org from scratch: `GOVERNMENT_ADMIN`,
    `PLATFORM_ADMIN`, `NGO`. Once created, the org's internal role
    (`OWNER` / `ADMIN` / `STAFF`) governs edits — not the platform JWT role.

    **Content-authorship writes are strictly org-only.** A `PLATFORM_ADMIN`
    who is not a member of the org **cannot** create, edit, publish, or
    delete announcements / projects / assignments / consultations on the
    org's behalf. Attribution belongs to the organization itself.

    Platform admins retain two narrow powers:

    - **Emergency close/archive** — `PLATFORM_ADMIN` can close a published
      consultation or archive an announcement without joining the org first.
      Written to the audit log with the actor's identity.
    - **Oversight reads** — `PLATFORM_ADMIN` can view responses, analytics,
      and org-only content (draft announcements, assignment inboxes) for
      any org.

    Anything beyond these two levers requires the platform admin to be
    added as a member of the target org first.
servers:
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ConsultationStatus:
      type: string
      enum: [DRAFT, PUBLISHED, CLOSED]
    QuestionType:
      type: string
      enum: [SHORT_TEXT, LONG_TEXT, SINGLE_CHOICE, MULTI_CHOICE, YES_NO]
...
paths:
...
  /consultations:
    get:
      tags: [Consultations]
      security: []
      summary: List consultations
      description: |
        Public list. Filter by organization, community, or status. Drafts
        are visible via the API but the citizen frontend hides them from
        non-members.
      parameters:
        - in: query
          name: organizationId
          schema: { type: string, format: uuid }
        - in: query
          name: communityId
          schema: { type: string, format: uuid }
        - in: query
          name: status
          schema: { $ref: '#/components/schemas/ConsultationStatus' }
        - in: query
          name: limit
          schema: { type: integer, default: 20 }
        - in: query
          name: offset
          schema: { type: integer, default: 0 }
      responses:
        '200':
          description: List of consultations.
          content:
            application/json:
              examples:
                default:
                  value:
                    success: true
                    data:
                      consultations:
                        - id: c1a2b3c4-5678-90ab-cdef-1234567890ab
                          organizationId: 9a8b7c6d-1111-2222-3333-444455556666
                          communityId: 6f2a1c4d-9b3e-4a1f-8c7d-2b3e4f5a6b7c
                          title: National Curriculum Review
                          summary: How should the syllabus change for 2027?
                          description: Full context in markdown.
                          status: PUBLISHED
                          responseCount: 142
                          authorId: 3f1a...
                          authorName: Ada Admin
                          publishedAt: 2026-07-09T10:00:00Z
                          createdAt: 2026-07-08T14:22:00Z
                          updatedAt: 2026-07-09T10:00:00Z
  /consultations/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [Consultations]
      security: []
      summary: Get a consultation by ID
      responses:
        '200':
          description: Consultation detail.
        '404':
          description: Not found.
          content:
            application/json:
              examples:
                default:
                  {
                    value:
                      {
                        success: false,
                        code: CONSULTATION_NOT_FOUND,
                        message: Consultation not found,
                      },
                  }
    patch:
      tags: [Consultations]
      summary: Edit a draft consultation (org admin)
      description: |
        Edits are DRAFT-only. Once published, the record is frozen so
        earlier and later responders answer the same questions.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title: { type: string, minLength: 5, maxLength: 200 }
                summary: { type: string, minLength: 10, maxLength: 500 }
                description: { type: string, minLength: 10 }
                coverImageUrl: { type: string, nullable: true }
                communityId: { type: string, format: uuid, nullable: true }
                opensAt: { type: string, format: date-time, nullable: true }
                closesAt: { type: string, format: date-time, nullable: true }
      responses:
        '200': { description: Updated. }
        '409':
          description: Not a draft.
          content:
            application/json:
              examples:
                default:
                  { value: { success: false, code: NOT_DRAFT, message: Only drafts can be edited } }
    delete:
      tags: [Consultations]
      summary: Delete a draft consultation (org admin)
      description: DRAFT-only. Once published, the record must survive.
      responses:
        '200': { description: Deleted. }
        '409': { description: Not a draft. }
  /organizations/{id}/consultations:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: string, format: uuid }
    post:
      tags: [Consultations]
      summary: Create a new consultation as DRAFT (org admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, summary, description]
              properties:
                title: { type: string, minLength: 5, maxLength: 200 }
                summary: { type: string, minLength: 10, maxLength: 500 }
                description: { type: string, minLength: 10 }
                coverImageUrl: { type: string, nullable: true }
                communityId: { type: string, format: uuid, nullable: true }
                opensAt: { type: string, format: date-time, nullable: true }
                closesAt: { type: string, format: date-time, nullable: true }
            examples:
              default:
                value:
                  title: National Curriculum Review
                  summary: How should the syllabus change for 2027?
                  description: We are collecting citizen input.
                  communityId: 6f2a1c4d-9b3e-4a1f-8c7d-2b3e4f5a6b7c
                  closesAt: 2026-09-30T23:59:59Z
      responses:
        '201':
          description: Created; status = DRAFT.
  /consultations/{id}/publish:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: string, format: uuid }
    post:
      tags: [Consultations]
      summary: Publish a draft consultation (org admin)
      description: DRAFT → PUBLISHED. Rejects if the consultation has no questions.
      responses:
        '200':
          description: Published; org members notified.
        '400':
          description: No questions yet.
          content:
            application/json:
              examples:
                default:
                  {
                    value:
                      {
                        success: false,
                        code: NO_QUESTIONS,
                        message: Add at least one question before publishing,
                      },
                  }
        '409':
          description: Invalid transition.
  /consultations/{id}/close:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: string, format: uuid }
    post:
      tags: [Consultations]
      summary: Close a published consultation (org admin OR platform admin)
      description: |
        PUBLISHED → CLOSED. Idempotent — closing a closed consultation is
        a no-op. Responders are notified so they can watch for the outcome.

        **Emergency lever.** This is the one consultation-lifecycle endpoint
        that a `PLATFORM_ADMIN` can invoke without being a member of the
        org. Every other write (create, edit, publish, delete, outcome) is
        strictly org-only. Platform-admin closes write an audit entry with
        the actor's identity.
      responses:
        '200': { description: Closed. }
        '409': { description: Invalid transition. }

  # ────────────────── CONSULTATION QUESTIONS ─────────────────
  /consultations/{id}/questions:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [Consultation Questions]
      security: []
      summary: List questions in order
      responses:
        '200':
          description: Questions.
          content:
            application/json:
              examples:
                default:
                  value:
                    success: true
                    data:
                      questions:
                        - id: q1a2b3c4-...
                          consultationId: c1a2b3c4-...
                          position: 1
                          prompt: Which subjects deserve more focus?
                          type: MULTI_CHOICE
                          options: [Maths, Science, History, Civic Ed]
                          required: true
    post:
      tags: [Consultation Questions]
      summary: Add a question to a draft (org admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [prompt, type]
              properties:
                prompt: { type: string, minLength: 3, maxLength: 500 }
                helpText: { type: string, nullable: true }
                type: { $ref: '#/components/schemas/QuestionType' }
                options:
                  type: array
                  items: { type: string }
                  description: Required for SINGLE_CHOICE and MULTI_CHOICE (min 2 items).
                required: { type: boolean, default: false }
                position: { type: integer, description: Omit to append at the end. }
      responses:
        '201': { description: Question created. }
        '400':
          description: Invalid type or missing options.
          content:
            application/json:
              examples:
                needsOptions:
                  {
                    value:
                      {
                        success: false,
                        code: OPTIONS_REQUIRED,
                        message: Choice questions need at least two options,
                      },
                  }
  /consultations/{id}/questions/reorder:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: string, format: uuid }
    patch:
      tags: [Consultation Questions]
      summary: Bulk-reorder questions (org admin, DRAFT only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ordering]
              properties:
                ordering:
                  type: object
                  additionalProperties: { type: integer }
                  description: Map of questionId → new zero-based position.
            examples:
              default:
                value:
                  ordering:
                    q1a2b3c4-...: 0
                    q5f6e7d8-...: 1
                    q9c8b7a6-...: 2
      responses:
        '200': { description: Reordered. }
  /consultation-questions/{questionId}:
    parameters:
      - in: path
        name: questionId
        required: true
        schema: { type: string, format: uuid }
    patch:
      tags: [Consultation Questions]
      summary: Update a question (org admin, DRAFT only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [prompt, type]
              properties:
                prompt: { type: string, minLength: 3, maxLength: 500 }
                helpText: { type: string, nullable: true }
                type: { $ref: '#/components/schemas/QuestionType' }
                options:
                  type: array
                  items: { type: string }
                required: { type: boolean }
                position: { type: integer }
      responses:
        '200': { description: Updated. }
    delete:
      tags: [Consultation Questions]
      summary: Delete a question (org admin, DRAFT only)
      responses:
        '200': { description: Deleted. }

  # ────────────────── CONSULTATION RESPONSES ─────────────────
  /consultations/{id}/responses:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: string, format: uuid }
    post:
      tags: [Consultation Responses]
      summary: Submit a response (verified user)
      description: |
        One response per user is enforced by a compound unique index on
        `(consultation_id, user_id)`. Community-scoped consultations do
        NOT gate submission — any verified user can respond regardless
        of community membership. The community label is informational.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [answers]
              properties:
                answers:
                  type: array
                  minItems: 1
                  items:
                    type: object
                    required: [questionId]
                    properties:
                      questionId: { type: string, format: uuid }
                      textValue:
                        type: string
                        nullable: true
                        description: For SHORT_TEXT and LONG_TEXT questions.
                      selections:
                        type: array
                        items: { type: string }
                        description: For SINGLE_CHOICE, MULTI_CHOICE, YES_NO questions. Yes/No is encoded as ["YES"] or ["NO"].
            examples:
              default:
                value:
                  answers:
                    - questionId: q1a2b3c4-...
                      selections: [YES]
                    - questionId: q2a2b3c4-...
                      selections: [Civic Ed, Science]
                    - questionId: q3a2b3c4-...
                      textValue: We should teach civics from primary school.
      responses:
        '201':
          description: Response recorded.
        '409':
          description: Consultation not accepting responses, closed, or already responded.
          content:
            application/json:
              examples:
                alreadyResponded:
                  {
                    value:
                      {
                        success: false,
                        code: ALREADY_RESPONDED,
                        message: You've already responded to this consultation,
                      },
                  }
                notAccepting:
                  {
                    value:
                      {
                        success: false,
                        code: NOT_ACCEPTING_RESPONSES,
                        message: This consultation isn't accepting responses,
                      },
                  }
                pastDeadline:
                  {
                    value:
                      { success: false, code: PAST_DEADLINE, message: Response window has closed },
                  }
        '400':
          description: Missing required answer, unknown question, or unknown option.
          content:
            application/json:
              examples:
                requiredUnanswered:
                  {
                    value:
                      {
                        success: false,
                        code: REQUIRED_UNANSWERED,
                        message: One or more required questions were not answered,
                      },
                  }
                unknownOption:
                  {
                    value:
                      {
                        success: false,
                        code: UNKNOWN_OPTION,
                        message: Selected option is not on the question,
                      },
                  }
    get:
      tags: [Consultation Responses]
      summary: List submitted responses (org admin)
      description: Members-only. Returns each response with its answers grouped.
      parameters:
        - in: query
          name: limit
          schema: { type: integer, default: 50 }
        - in: query
          name: offset
          schema: { type: integer, default: 0 }
      responses:
        '200':
          description: Responses.
...